Use (long) for Dynamodb _setcapacity()
The code which parses ReadCapacityUnits and WriteCapacityUnits uses a
(long), but that's truncated to an (int) when they're passed to
dynamodb_request_queue_setcapacity().  Internally, that function sets
Q->cappers which is a (double), so there's no harm changing the argument.

It's unlikely that we'll want to do more than INT_MAX operations per
second, of course.

+++ lib/dynamodb/dynamodb_request_queue.h
@@ -33,7 +33,7 @@ void dynamodb_request_queue_log(struct dynamodb_request_queue *,
  * Exceeded" exception.  If passed a capacity of 0, the request rate will
  * not be limited.
  */
-void dynamodb_request_queue_setcapacity(struct dynamodb_request_queue *, int);
+void dynamodb_request_queue_setcapacity(struct dynamodb_request_queue *, long);
 
 /**
  * dynamodb_request_queue(Q, prio, op, body, maxrlen, logstr, callback, cookie):
